refactor(gda): move the live-stack groups into gda.commands (ADR-0040)
game, diag, logger, perf, input, screen, and daemon each own their
vertical slice; screen_ops.py and daemon_ops.py are absorbed by their
groups; cli.py is reduced to the composition root (root app, callback,
register calls). Surface manifest verified against the pre-slice state.
